Add support for disabling form action buttons (copy, edit, remove)
In my case there is a need for static fields that users are not allowed to remove or edit. Therefore I added support this by allowing developers to specify in the field values, through a disableFieldButtons array containing one or more of ['copy', 'edit', 'remove'].

Whenever disabling the edit field button, the dblclick event on the field will also be ignored.